# specs/001-ai-sop-mvp/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: AI SOP MVP API
  version: 0.1.0
  description: REST API for conversational SOP generation, PDCA management, and weekly reviews.
servers:
  - url: https://api.example.com/api/v1
    description: Production
  - url: http://localhost:4000/api/v1
    description: Development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SopId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    ActionCard:
      type: object
      required: [id, title, description, ownerId, status]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        ownerId:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, in_review, done, blocked]
        dueAt:
          type: string
          format: date-time
        tags:
          type: object
          additionalProperties: true
    SOP:
      type: object
      required: [id, title, status, tags]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        projectId:
          type: string
          format: uuid
        status:
          type: string
          enum: [draft, in_review, approved, archived]
        tags:
          type: object
          properties:
            domain:
              type: string
            role:
              type: string
            phase:
              type: string
            tool:
              type: string
            risk:
              type: string
            scope:
              type: string
            free:
              type: array
              items:
                type: string
        notes:
          type: string
        currentVersion:
          $ref: '#/components/schemas/SOPVersion'
    SOPVersion:
      type: object
      required: [id, sopId, semver, status]
      properties:
        id:
          type: string
          format: uuid
        sopId:
          type: string
          format: uuid
        semver:
          type: string
        status:
          type: string
          enum: [draft, pending_review, approved, superseded]
        contentMarkdown:
          type: string
        contentJson:
          type: object
        diffJson:
          type: object
        changeSummary:
          type: string
        reviewedBy:
          type: string
          format: uuid
        reviewedAt:
          type: string
          format: date-time
...
security:
  - bearerAuth: []
paths:
...
  /sops:
    get:
      tags: [SOPs]
      summary: List SOPs with filtering and pagination.
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: tag.domain
          schema:
            type: string
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: SOP list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/SOP'
                  total:
                    type: integer
    post:
      tags: [SOPs]
      summary: Create SOP draft from action cards.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, actions]
              properties:
                title:
                  type: string
                projectId:
                  type: string
                  format: uuid
                actions:
                  type: array
                  items:
                    $ref: '#/components/schemas/ActionCard'
                tags:
                  type: object
      responses:
        '201':
          description: SOP draft created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SOP'
  /sops/{id}:
    get:
      tags: [SOPs]
      summary: Get SOP detail with latest version.
      parameters:
        - $ref: '#/components/parameters/SopId'
      responses:
        '200':
          description: SOP detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SOP'
    patch:
      tags: [SOPs]
      summary: Update SOP draft metadata or actions.
      parameters:
        - $ref: '#/components/parameters/SopId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                tags:
                  type: object
                notes:
                  type: string
                actions:
                  type: array
                  items:
                    $ref: '#/components/schemas/ActionCard'
      responses:
        '200':
          description: SOP updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SOP'
...